feat: Add and refactor Vue-based "Motores" shop page
This commit introduces a new public-facing "Motores" shop page and refactors it to align with existing project patterns.

The new page is located at `/#/tienda` and includes:
- A card-based layout to display motors.
- Reactive filtering for category, brand, country, and item condition.
- Pagination.
- A navigation link in the main menu.

The implementation was refactored to use the reactive API helpers (`useApi`, `createUrl`) found in the existing codebase, making the data fetching more efficient and consistent. The file structure was also reorganized for better clarity.

This change also resolves a runtime error caused by a faulty router link and i18n warnings from the initial implementation.
